requirements.txt
appdirs==1.4.4
audioread==2.1.9
autopep8==1.5.6
boto3==1.17.46
botocore==1.20.46
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==5.0.5
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
librosa==0.8.0
llvmlite==0.36.0
MarkupSafe==1.1.1
numba==0.53.1
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pooch==1.3.0
pycodestyle==2.7.0
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
resampy==0.2.2
s3transfer==0.3.6
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
